refactor: adjust visibility, move file
